server/django_backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
cryptography==43.0.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
etelemetry==0.3.1
filelock==3.15.4
filetype==1.2.0
fsspec==2024.6.1
ftfy==6.2.0
grpcio==1.65.1
httplib2==0.22.0
huggingface-hub==0.24.2
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
looseversion==1.3.0
lxml==5.2.2
marker-pdf==0.2.16
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdfminer.six==20240706
pdftext==0.3.10
pillow==10.4.0
prov==2.0.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydot==3.0.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rapidfuzz==3.9.4
rdflib==6.3.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.14.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.1
surya-ocr==0.4.15
sympy==1.13.1
tabulate==0.9.0
texify==0.1.10
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
traits==6.3.2
transformers==4.43.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
